# src/main/java/org/example/bankcards/dto/response/ErrorResponse.java
package org.example.bankcards.dto.response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ErrorResponse {

    private String error;

    private String message;
}